# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HCIS YSQ API
  version: 0.0.1
  description: |
    Kontrak awal untuk vertical slice identity, employee profile, leave, organization, dan attendance operations.
    Status dokumen DRAFT kecuali kontrak AUTH-011 /auth/me yang mengikuti implementasi.
    AUTH-011 admin route permissions: docs/domain/AUTH-011-admin-route-permissions.json.
    Every administrative operation requires the declared effective organization permission;
    deprecated legacy HCIS SUPER_ADMIN compatibility grants no platform or workflow approval authority.
servers:
  - url: https://api.example.invalid
    description: Placeholder only
...
paths:
...
  /admin/attendance/adms/management/summary:
    get:
      tags: [Attendance]
      operationId: getAdmsManagementSummary
      summary: Read fleet-level ADMS operational summary
      responses:
        '200': { description: Fleet, unmapped PIN, review, and command counters }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/attendance/adms/management/user-sync:
    get:
      tags: [Attendance]
      operationId: listAdmsUserSynchronization
      summary: List desired-versus-observed ADMS user state across devices
      parameters:
        - { name: deviceId, in: query, schema: { type: string, format: uuid } }
        - { name: state, in: query, schema: { type: string, enum: [synced, unmapped, missing_on_device, name_drift, review_required] } }
        - { name: q, in: query, schema: { type: string, maxLength: 100 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 250, default: 100 } }
      responses:
        '200': { description: Safe fleet-level user synchronization rows; no biometric payload }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/attendance/adms/management/jobs:
    get:
      tags: [Attendance]
      operationId: listAdmsManagementJobs
      summary: List human-readable ADMS command lifecycle across devices
      parameters:
        - { name: deviceId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string, enum: [pending, delivered, acknowledged, succeeded, failed, expired, cancelled] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 250, default: 100 } }
      responses:
        '200': { description: Command/job summaries without wire commands }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /admin/attendance/adms/management/audit:
    get:
      tags: [Attendance]
      operationId: listAdmsManagementAudit
      summary: List safe append-only ADMS administration audit summaries
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 250, default: 100 } }
      responses:
        '200': { description: Audit summaries without raw request or biometric payloads }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /admin/attendance/adms/devices/{deviceId}/retire:
    parameters:
      - name: deviceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Attendance]
      operationId: retireAdmsDevice
      summary: Retire an ADMS device while preserving historical evidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note: { type: string, minLength: 5, maxLength: 500 }
                replacementDeviceId: { type: [string, 'null'], format: uuid }
      responses:
        '200': { description: Retired device with optional replacement reference }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Device already retired or replacement invalid }
...
components:
...
  responses:
    Unauthenticated:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Actor is not authorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found or intentionally concealed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        correlationId:
          type: string
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string